Refactor Step 3
More moving to different packages, allowing git to track easily.
