+++ src/main/resources/com/insightfullogic/honest_profiler/ports/javafx/FlameGraphView.fxml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import com.insightfullogic.honest_profiler.ports.javafx.flame_graph.FlameGraphCanvas?>
+<?import com.insightfullogic.honest_profiler.ports.javafx.view.FlameGraphCanvas?>
 <FlameGraphCanvas fx:controller="com.insightfullogic.honest_profiler.ports.javafx.controller.FlameViewController"
                   fx:id="flameView"
                   xmlns="http://javafx.com/javafx/8"
